+++ src/com/intellij/idea/plugin/hybris/runConfigurations/HybrisJUnitExtension.java
@@ -1,7 +1,7 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -62,7 +62,7 @@ public class HybrisJUnitExtension extends RunConfigurationExtension {
             }
         }
 
-        if (!params.getEnv().containsKey(HYBRIS_DATA_DIR_ENV)) {
+        if (!params.getEnv().containsKey(ENV_HYBRIS_DATA_DIR)) {
             final HybrisProjectSettings settings =
                 HybrisProjectSettingsComponent.getInstance(project).getState();
 
@@ -75,7 +75,7 @@ public class HybrisJUnitExtension extends RunConfigurationExtension {
                         + HYBRIS_DATA_DIRECTORY);
 
             if (hybrisDataDirPath != null) {
-                params.addEnv(HYBRIS_DATA_DIR_ENV, hybrisDataDirPath);
+                params.addEnv(ENV_HYBRIS_DATA_DIR, hybrisDataDirPath);
             }
         }